refactor: refactor cjsToEsm
